// packages/backend-canvas2d/src/generated.ts
export const ABI_VERSION = 23 as const;
export const MINIMUM_READABLE_ABI_VERSION = 4 as const;
export const PROTOCOL_ALIGNMENT = 4 as const;
export const STREAM_HEADER_BYTES = 16 as const;
// ...
export const DISPLAY_LIST_MAGIC = 1146113860 as const;
// ...
export enum DisplayOpcode {
  Save = 1,
  Restore = 2,
  Transform = 3,
  ClipRect = 4,
  Alpha = 5,
  FillRect = 16,
  FillRRect = 17,
  FillPath = 18,
  StrokePath = 19,
  DrawGlyphRun = 32,
  DrawTextFallback = 33,
  DrawTextInlineFallback = 36,
  FillPlaceholder = 38,
  FillColorRect = 39,
  FillColorRRect = 40,
  FillColorBorder = 41,
  DrawEditorDecoration = 37,
  FillColorShadow = 42,
  FillColorPath = 43,
  StrokeColorPath = 44,
  DrawImage = 34,
  DrawPicture = 35,
}